Add the UK federation registration import validator
This is something of a brute-force first attempt at the problem,
essentially dropping in everything that is required without regard to
whether it should be in here or in the parent container.

This is particularly true for the binary dependencies: those should
almost all be pulled in as explicit dependencies in the parent
container rather than held as binary resources here.

We will need to return to all of this once we have the proof of
concept working.